About page resolved
